start cluster when adress is specified without a cluster
